add build-time option for setting any pixel format
this allows for setting 8,16,32 bit pixels, and allowing the user to
set a ptr into the pixel buffer themselves.

this will be a nice speedup for software rendering, as pixels can be
written directly to the screen

furthermore, this is useful for systems with strange display setups
such as one embeded system i worked on used 256 LUT for screen pixels
which actually just indexed into a colour array, so the LUT acted as
colour palette table.

this is not yet finished, gbc support needs to be added by either
adding a callback on palette swap, so the user can mix the colours
themelves, or the core can handle it.

i think the callback is better suited as it's really out of scope for a
gb core to be in charge of colour mixing...

+++ src/tables/palette_table.h
@@ -28,14 +28,14 @@ enum GB_CustomPalette {
 };
 
 struct GB_PaletteEntry {
-    uint16_t BG[4];
-    uint16_t OBJ0[4];
-    uint16_t OBJ1[4];
+    uint32_t BG[4];
+    uint32_t OBJ0[4];
+    uint32_t OBJ1[4];
 };
 
 struct GB_PalettePreviewShades {
-    uint16_t shade1;
-    uint16_t shade2;
+    uint32_t shade1;
+    uint32_t shade2;
 };
 
 /* RETURNS: */
